refactor(benchmarks-website): split web_ui.rs into topical test files
Pull the 23 integration tests in server/tests/web_ui.rs into five
topical test crates plus a shared `common/mod.rs` fixture module:

- common/mod.rs        — Server harness, seed/seed_long_history,
  envelope_for/ra_envelope_for, extract_chart_data, pick_chart_slug,
  pick_group_slug, group_by_name, assert_close, filter_bar_section,
  filter_section, extract_chip, insta_settings, TOKEN.
- landing.rs           — landing page + global filter chip tests.
- chart_api.rs         — `/chart/{slug}` and `/api/chart/{slug}` tests.
- group_api.rs         — `/group/{slug}`, `/api/group/{slug}`, group
  summary contract tests.
- permalinks.rs        — full-history inline payload, embedded filter
  state, unknown-slug 404 tests.
- static_assets.rs     — bundled static asset routes + compression
  layer tests.

Snapshot files don't move because every `assert_snapshot!` call uses
`set_prepend_module_to_snapshot(false)` plus an explicit name, so the
snapshot key is path-independent. The four existing snapshot files
in `server/tests/snapshots/` continue to match byte-for-byte
without `INSTA_UPDATE`.
